Update INSTALL.md
I don't know if it's just me but it seems that Ninja can't detect C:\deps\depot_tools\python.bat and needs PATH to explicitly point to python.exe. Not tested, (kind of hard to isolate and test just this on my extremely cluttered PC, sorry) so should be tested before merging is considered or anything like that.